AFGS-525 Letter payment amounts are now formatted the same way as they are in emails.
